repo-tools: refactor sql-reports into split modules
